# swagger.yaml
{"openapi":"3.0.0","info":{"title":"Trentino Eventi","description":"Applicazione che mostra eventi ed alloggi in Trentino","version":"1.0.0"},"servers":{"url":{"http":"//localhost:8080/","description":"Localhost"}},"paths":{"/api/v1/authentication/login":{"post":{"description":"Controlla se l utente esiste e in caso affermativo fa il login","summary":"Fa il login se l utente esiste","tags":["authentication"],"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"type":"string","description":"Indirizzo email dell utente"},"password":{"type":"string","description":"Password dell utente"}}}}}},"responses":{"200":{"description":"Utente esiste, aggiunge il cookie contentente il token, il nome dell utente e l id alla response o password errata restituisce messaggio","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale true se non ci sono stati errori, false se la password è sbagliata"}}}}}},"400":{"description":"Restituisce errore se non sono stati inseriti tutti i campi!","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale false"},"message":{"type":"string","description":"Messaggio che contiene l'errore"}}}}}},"404":{"description":"Restituisce errore utente non trovato!","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale false ed indica che ci sono stati errori"},"message":{"type":"string","description":"Messaggio che contiene l'errore"}}}}}}}}},"/api/v1/authentication/subscribe":{"post":{"description":"Controlla se l utente esiste e in caso non affermativo lo iscrive, aggiunge il cookie contentente il token, il nome dell utente e l id alla response","summary":"Fa la registazione se l utente non esiste","tags":["authentication"],"requestBody":{"content":{"application/json":{"schema":{"properties":{"name":{"type":"string","description":"Nome dell utente"},"surname":{"type":"string","description":"Cognome dell utente"},"birthdate":{"type":"string","description":"Data di nascita dell utente"},"userType":{"type":"string","description":"Tipo di utente"},"email":{"type":"string","description":"Indirizzo email dell utente"},"password":{"type":"string","description":"Password di utente"}}}}}},"responses":{"200":{"description":"Registrazione avvenuta correttamente","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale true ed indica che non ci sono stati errori"}}}}}},"400":{"description":"Restituisce errori in caso di campi non inseriti, formato data o email non corretto, o tipo di utente non corretto","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale false"},"message":{"type":"string","description":"Messaggio che contiene l'errore"}}}}}},"404":{"description":"Restituisce errore utente gia esistente!","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale false ed indica che ci sono stati errori"},"message":{"type":"string","description":"Messaggio che contiene l'errore"}}}}}}}}},"/api/v1/authentication/checkIfLogged":{"get":{"description":"Controlla se l utente e gia loggato","summary":"Controlla se utente loggato","tags":["authentication"],"responses":{"200":{"description":"Utente gia loggato","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale true ed indica che non ci sono stati errori"},"name":{"type":"string","description":"Contiene il nome dell utente"},"message":{"type":"string","description":"Messaggio che contiene un messaggio di successo"}}}}}},"404":{"description":"Restituisce errori cookie non trovato e token non valido","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale false ed indica che ci sono stati errori"},"message":{"type":"string","description":"Messaggio che contiene l'errore"}}}}}}}}},"/api/v1/authentication/logout":{"get":{"description":"Cancella il cookie dell utente se esiste","summary":"Fa il logout dell utente","tags":["authentication"],"responses":{"200":{"description":"Restituisce un messaggio che dice che il cookie non e stato trovato o che e stato cancellato correttamente a seconda che lo trovi o meno","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"Vale true se il cookie e stato cancellato e false se non esisteva"},"message":{"type":"string","description":"Messaggio che contiene informazioni sull azione fatta"}}}}}}}}},"/api/v1/visualizzazione/eventList":{"get":{"description":"Gets the list of all events","summary":"View all events","tags":["eventVisualization"],"responses":{"200":{"description":"Collection of events in JSON format","content":{"application/json":{"schema":{"properties":{"id":{"type":"string","description":"string that identifies the event"},"title":{"type":"string","description":"title of the event"}}}}}}}}},"/api/v1/visualizzazione/housingList":{"get":{"description":"Gets the list of all housings","summary":"View all housings","tags":["housingVisualization"],"responses":{"200":{"description":"Collection of housings in JSON format","content":{"application/json":{"schema":{"properties":{"id":{"type":"string","description":"string that identifies the housing"},"title":{"type":"string","description":"title of the housing"}}}}}}}}},"/api/v1/visualizzazione/event":{"get":{"description":"Gets the details of a specific event","summary":"Details of one event","tags":["eventVisualization"],"parameters":[{"in":"query","name":"id","type":"string","description":"The id of the specific event you want to get the details","required":true}],"responses":{"200":{"description":"Details of an event in JSON format","content":{"application/json":{"schema":{"properties":{"title":{"type":"string","description":"title of the event"},"description":{"type":"string","description":"description of the event"},"initDate":{"type":"string","description":"date of the beginning of the event in ISO8601 format"},"finlDate":{"type":"string","description":"date of the ending of the event in ISO8601 format"},"address":{"type":"string","description":"address of the event"},"city":{"type":"string","description":"city where the event takes place"},"seatsAvailable":{"type":"integer","description":"seats still available for the event"},"seatsOccupied":{"type":"number","description":"seats already occupied for the event"},"category":{"type":"number","description":"seats already occupied for the event"},"creatorName":{"type":"string","description":"the name of the user who created the event"},"creatorSurname":{"type":"string","description":"the surname of the user who created the event"},"creatorEmail":{"type":"string","description":"the email of the user who created the event"}}}}}},"400":{"description":"The Id query parameter is not present or do not respect the MongoDB format","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"specifications regarding the specific error"}}}}}},"404":{"description":"Some data are not found on the DB","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"specifications regarding the specific error"}}}}}}}}},"/api/v1/visualizzazione/housing":{"get":{"description":"Gets the details of a specific housing","summary":"Details of one housing","tags":["housingVisualization"],"parameters":[{"in":"query","name":"id","type":"string","description":"The id of the specific housing you want to get the details","required":true}],"responses":{"200":{"description":"Details of a housing in JSON format","content":{"application/json":{"schema":{"properties":{"title":{"type":"string","description":"title of the housing"},"description":{"type":"string","description":"description of the housing"},"initDate":{"type":"string","description":"date of the beginning of the housing availability slot in ISO8601 format"},"finlDate":{"type":"string","description":"date of the ending of the housing availability slot in ISO8601 format"},"address":{"type":"string","description":"address of the housing"},"city":{"type":"string","description":"city where the housing is located"},"creatorName":{"type":"string","description":"the name of the user who created the housing"},"creatorSurname":{"type":"string","description":"the surname of the user who created the housing"},"creatorEmail":{"type":"string","description":"the email of the user who created the housing"}}}}}},"400":{"description":"The Id query parameter is not present or do not respect the MongoDB format","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"specifications regarding the specific error"}}}}}},"404":{"description":"Some data are not found on the DB","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"specifications regarding the specific error specific error"}}}}}}}}},"/api/v1/eventSubscription/eventSubcribable":{"post":{"description":"Check if the user is alreadi subscribed to the specific event","summary":"Check user subscription to event","tags":["eventSubscription"],"parameters":[{"in":"body","name":"token","type":"string","description":"The token that rapresent the logged user","required":true},{"in":"body","name":"event","type":"string","description":"The id of the event","required":true}],"responses":{"200":{"description":"Request successful","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"UserNotSubscribed => the user is not subscribed to the specific event\n\nUserSubscribed => the user is already subscribed to the specific event\n"}}}}}},"401":{"description":"The user is not logged","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean"},"message":{"type":"string","description":"UserNotLogged => the user has not provided a valid token, therefore the user is not logged\n"}}}}}}}}},"/api/v1/eventSubscription/createSubscription":{"post":{"description":"Require to subscrive the user to the specific event","summary":"Subscribe the user to the event","tags":["eventSubscription"],"parameters":[{"in":"body","name":"token","type":"string","description":"The token that rapresent the logged user","required":true},{"in":"body","name":"event","type":"string","description":"The id of the event to subscribe to","required":true}],"responses":{"200":{"description":"The request was successful but it is not guaranteed that the new entry has been created, check the message field","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"NoFreeSeats => false\n\nUserAlreadySubscribed => true\n"},"message":{"type":"string","description":"NoFreeSeats => the event has no free seats, the subscription has not been created\n\nUserAlreadySubscribed => the user is already subscribed to the specific event\n"}}}}}},"201":{"description":"The request was successful and the new entry has been created","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"UserSubscribed => true\n"},"message":{"type":"string","description":"UserSubscribed => the user has been subscribed to the event correctly and the new entry has been created\n"}}}}}},"401":{"description":"The user is not logged","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","description":"UserNotLogged => false\n"},"message":{"type":"string","description":"UserNotLogged => the user has not provided a valid token, therefore the user is not logged\n"}}}}}}}}}},"components":{},"tags":[]}